File changes in errors.py and libhash.py
-rename the file libhash.py to lib.py
-transfer the "error_catching" function to lib.py
